src/python/OCR_Vizwiz/requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
tqdm
Pillow
matplotlib
easyocr
